Working into support ipynb with different kernels. Right now splitting things, but still WIP.
