+++ nikola/plugins/compile/ipynb_python2.plugin
@@ -1,6 +1,6 @@
 [Core]
-Name = ipynb
-Module = ipynb/ipynb
+Name = ipynb_python2
+Module = ipynb/ipynb_python2
 
 [Documentation]
 Author = Damian Avila
